pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "likedmusic"
version = "0.1.0"
description = "Sync YouTube Music liked songs to Apple Music"
requires-python = ">=3.10"
dependencies = [
    "ytmusicapi>=1.11.0",
    "yt-dlp",
    "mutagen",
    "requests",
    "rookiepy",
    "pyyaml",
    "questionary>=2.0.0",
    "rich>=13.0.0",
]

[project.optional-dependencies]
dev = ["pytest"]

[project.scripts]
likedmusic = "likedmusic.cli:main"

[tool.pytest.ini_options]
testpaths = ["tests"]
